feat!: api usage and config separation
